Add new entities for modules
