+++ src/domain/common/entities/NamedRef.ts
@@ -1,4 +1,4 @@
-import { Ref } from "./Schemas";
+import { Ref } from "./Ref";
 
 export interface NamedRef extends Ref {
     name: string;
