+++ src/domain/common/entities/Ref.ts
@@ -1,0 +1,5 @@
+import { Id } from "./Schemas";
+
+export interface Ref {
+    id: Id;
+}
