+++ src/domain/metadata/usecases/MetadataSyncUseCase.ts
@@ -6,7 +6,7 @@ import { D2 } from "../../../types/d2";
 import { D2Api } from "../../../types/d2-api";
 import { ExportBuilder, NestedRules, SynchronizationBuilder } from "../../../types/synchronization";
 import { promiseMap } from "../../../utils/common";
-import { Ref } from "../../common/entities/Schemas";
+import { Ref } from "../../common/entities/Ref";
 import InstanceEntity from "../../instance/entities/Instance";
 import InstanceRepository from "../../instance/repositories/InstanceRepository";
 import { SynchronizationResult } from "../../synchronization/entities/SynchronizationResult";
