+++ src/domain/modules/entities/Snapshot.ts
@@ -1,0 +1,12 @@
+import { NamedRef } from "../../common/entities/NamedRef";
+import { MetadataPackage } from "../../metadata/entities/MetadataEntities";
+
+export interface Snapshot extends NamedRef {
+    module: string;
+    revision: string;
+    author: {
+        name: string;
+        email: string;
+    };
+    contents: MetadataPackage;
+}
