+++ src/domain/modules/entities/Store.ts
@@ -1,0 +1,8 @@
+// Represents a git repository that stores modules
+export interface Store {
+    token: string;
+    account: string;
+    repository: string;
+    branch?: string;
+    basePath?: string;
+}
