+++ src/domain/synchronization/entities/SynchronizationRule.ts
@@ -1,23 +1,14 @@
 import { SynchronizationBuilder } from "../../../types/synchronization";
-import { NamedRef } from "../../common/entities/NamedRef";
-import { SharingSetting } from "../../common/entities/SharingSetting";
+import { SharedObject } from "../../common/entities/SharedObject";
 
-export interface SynchronizationRule {
-    id: string;
-    name: string;
+export interface SynchronizationRule extends SharedObject {
     code?: string;
     created: Date;
     description?: string;
     builder: SynchronizationBuilder;
     enabled: boolean;
     lastExecuted?: Date;
-    lastUpdated: Date;
-    lastUpdatedBy: NamedRef;
     frequency?: string;
-    publicAccess: string;
-    user: NamedRef;
-    userAccesses: SharingSetting[];
-    userGroupAccesses: SharingSetting[];
     type: SyncRuleType;
 }
